backend/requirements.txt
fastapi==0.115.12
python-dotenv==1.1.0
strands-agents[ollama]==1.6.0
uvicorn[standard]==0.34.2
